ENH: added capability to parse multi frame data by TemporalPositionIdentifier.
Units are currently set to "count", as in the particular type of series that motivated
this use case, 2D acquisition was used and each slice had a distinctive
TriggerTime. In the future, it might make sense to take the time-stamp of the
first slice in each frame and assign it as the time-stamp for the frame,
however it would break the uniformity of processing, and it is not clear if
this is the best approach for temporal analysis.