ENH: Allow loading of multi-volumes when multiple series are selected in the DICOM browser
examineFilesIPPAcqTime fails to recognize 4D Cardiac CT data sets (acquired on Siemens Somatom Definition CT) if multiple series are selected in the DICOM browser.
The problem is that if all selected series are passed to examineFilesIPPAcqTime then the extra series cause varying slice number for each IPP value,
so nSlicesEqual becomes false.

The solution is to pass series to examineFilesIPPAcqTime one-by-one.