ENH: change the confidence for the loadables
Loadables were always assigned confidence 1. It was also the case that loadables recognized
by TemporalPositionIdentifier were preferred over another strategy, while not being usable for
DSC analysis for a sample dataset provided by a user.

Now multivolumes are assigned confidence 0.9 except when TemporalPositionIdentifier is used, which
is assigned 0.8 so that default choice is more usable for quantitative analysis.