ENH: Improve logging during multi-volume import
When application settings / DICOM / detailed logging checkbox is checked then the plugin logs detailed information about why a series is not accepted as multi-volume.
The information is not logged by default because it would add too much noise to the log and users may interpret the diagnostic information as errors.