Transfer attributes to Sequence node
A number of attributes are set on MultiVolume nodes on import, including "MultiVolume.DICOM.EchoTime", "MultiVolume.DICOM.FlipAngle", "MultiVolume.DICOM.RepetitionTime", "MultiVolume.FrameIdentifyingDICOMTagName", and "MultiVolume.FrameLabels".  These were used in modules such as DSCMRIAnalysis.  However, they were not transferred to the volume Sequence node if a Sequence node was imported instead of a MultiVolume node.  Now that the MultiVolume selector widget has been modified to allow selection of Sequence nodes, these attributes need to be transferred in order for things to work the same way as they did for MultiVolumes.  This commit transfers all attributes stored for a MultiVolume node to the corresponding Sequence node on import.